refactor: move cloud base url validation to its own file
