+++ rust/alera-cli/src/alera_account/cloud_client.rs
@@ -1,10 +1,11 @@
-use anyhow::{anyhow, Context as _, Result};
+use anyhow::{Context as _, Result};
 use chrono::{DateTime, Utc};
 use reqwest::{Client, Method, StatusCode};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use serde_json::Value;
 use thiserror::Error;
-use url::{Host, Url};
+
+use crate::terminal_host::alera_account::cloud_base_url::validate_cloud_base_url;
 
 pub(crate) const DEFAULT_CLOUD_BASE_URL: &str = "https://api.alera.build";
 
@@ -394,27 +395,6 @@ impl CloudAccountClient {
     }
 }
 
-/// Cleartext HTTP is accepted only for a literal loopback origin. A prefix match is
-/// not enough: `http://localhost.example.com` starts with `http://localhost` while
-/// resolving to a remote host, which would put bearer tokens on the wire in the clear.
-pub(crate) fn validate_cloud_base_url(base_url: &str) -> Result<()> {
-    let parsed =
-        Url::parse(base_url).map_err(|_| anyhow!("ALERA_CLOUD_URL must be an absolute URL"))?;
-    let loopback = match parsed.host() {
-        Some(Host::Domain(domain)) => domain.eq_ignore_ascii_case("localhost"),
-        Some(Host::Ipv4(address)) => address.is_loopback(),
-        Some(Host::Ipv6(address)) => address.is_loopback(),
-        None => false,
-    };
-    match parsed.scheme() {
-        "https" => Ok(()),
-        "http" if loopback => Ok(()),
-        _ => Err(anyhow!(
-            "ALERA_CLOUD_URL must use HTTPS or a loopback HTTP origin"
-        )),
-    }
-}
-
 fn cloud_error(status: StatusCode, body: &str, path: &str) -> anyhow::Error {
     let parsed = serde_json::from_str::<Value>(body).ok();
     let error = parsed.as_ref().and_then(|value| value.get("error"));
@@ -441,38 +421,7 @@ fn cloud_error(status: StatusCode, body: &str, path: &str) -> anyhow::Error {
 mod tests {
     use reqwest::StatusCode;
 
-    use super::{cloud_error, validate_cloud_base_url, CloudRequestError};
-
-    #[test]
-    fn loopback_lookalike_hosts_are_rejected() {
-        for base_url in [
-            "http://localhost.example.com",
-            "http://127.0.0.1.example.com",
-            "http://api.alera.build",
-            "ftp://localhost",
-            "not a url",
-        ] {
-            assert!(
-                validate_cloud_base_url(base_url).is_err(),
-                "{base_url} must be rejected"
-            );
-        }
-    }
-
-    #[test]
-    fn https_and_real_loopback_origins_are_accepted() {
-        for base_url in [
-            "https://api.alera.build",
-            "http://localhost:8787",
-            "http://127.0.0.1:8787",
-            "http://[::1]:8787",
-        ] {
-            assert!(
-                validate_cloud_base_url(base_url).is_ok(),
-                "{base_url} must be accepted"
-            );
-        }
-    }
+    use super::{cloud_error, CloudRequestError};
 
     #[test]
     fn relay_key_conflicts_remain_machine_readable() {
