+++ rust/alera-cli/src/alera_account/mod.rs
@@ -1,12 +1,12 @@
+pub(crate) mod cloud_base_url;
 mod cloud_client;
 mod credential_store;
 mod loopback_callback;
 mod pkce;
 mod service;
 
-pub(crate) use cloud_client::{
-    validate_cloud_base_url, AuthProvider, AuthTransaction, PushEventRequest, RelayGrant,
-};
+pub(crate) use cloud_base_url::validate_cloud_base_url;
+pub(crate) use cloud_client::{AuthProvider, AuthTransaction, PushEventRequest, RelayGrant};
 pub(crate) use loopback_callback::{bind_callback_listener, wait_for_callback};
 pub(crate) use pkce::Pkce;
 pub(crate) use service::AleraAccountService;
